requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
databases==0.5.5
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.0
orjson==3.6.7
passlib==1.7.4
psycopg-binary==3.0.10
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2
zipp==3.7.0
